+++ nubix/__init__.py
@@ -1,5 +1,5 @@
 """Nubix - Cloud Sync Manager for Ubuntu."""
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 __author__ = "Nubix Contributors"
 __license__ = "Proprietary"